# docs/swagger.yaml
openapi: 3.0.3
info:
  title: UnaHur Anti-Social Net API
  description: |
    API REST del backend para la red social "UnaHur Anti-Social Net".
    Permite gestionar usuarios, publicaciones, imágenes, etiquetas y comentarios.
  version: 1.0.0
servers:
  - url: http://localhost:3000
    description: Servidor local de desarrollo
...
paths:
...
  /v1/posts/{post_id}/comments:
    parameters:
      - name: post_id
        in: path
        required: true
        description: ID del post
        schema:
          type: integer
          minimum: 1
    get:
      tags: [Comentarios]
      summary: Listar comentarios de un post
      description: Incluye datos del usuario autor. Ordenados por fecha de creación descendente.
      operationId: getCommentsByPost
      responses:
        '200':
          description: Lista de comentarios
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CommentWithUser'
        '404':
          description: Post no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: Post no encontrado
        '500':
          $ref: '#/components/responses/ServerError'
    post:
      tags: [Comentarios]
      summary: Crear un comentario en un post
      operationId: postCommentByPost
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CommentInput'
      responses:
        '201':
          description: Comentario creado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Comment'
        '400':
          description: Error de validación
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/CommentValidationError'
                  - $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Post o usuario no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/ServerError'
  /v1/posts/{post_id}/comments/{comment_id}:
    parameters:
      - name: post_id
        in: path
        required: true
        description: ID del post
        schema:
          type: integer
          minimum: 1
      - name: comment_id
        in: path
        required: true
        description: ID del comentario
        schema:
          type: integer
          minimum: 1
    put:
      tags: [Comentarios]
      summary: Actualizar un comentario
      operationId: putCommentByPost
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CommentUpdateInput'
      responses:
        '200':
          description: Comentario actualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Comment'
        '400':
          description: Error de validación
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/CommentValidationError'
                  - $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Post o comentario no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/ServerError'
    delete:
      tags: [Comentarios]
      summary: Eliminar un comentario
      operationId: deleteCommentByPost
      responses:
        '200':
          description: Comentario eliminado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '404':
          description: Post o comentario no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/ServerError'
components:
...
  schemas:
...
    Comment:
      type: object
      properties:
        idComment:
          type: integer
        idPost:
          type: integer
        idUser:
          type: integer
        contenido:
          type: string
          maxLength: 255
        visible:
          type: boolean
          description: Campo virtual; false si el comentario supera los 60 días
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CommentWithUser:
      allOf:
        - $ref: '#/components/schemas/Comment'
        - type: object
          properties:
            User:
              type: object
              properties:
                idUser:
                  type: integer
                nickName:
                  type: string
                nombre:
                  type: string
                apellido:
                  type: string
    CommentInput:
      type: object
      required: [idUser, contenido]
      properties:
        idUser:
          type: integer
          minimum: 1
        contenido:
          type: string
          minLength: 1
          maxLength: 255
    CommentUpdateInput:
      type: object
      required: [contenido]
      properties:
        contenido:
          type: string
          minLength: 1
          maxLength: 255
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
...
    CommentValidationError:
      type: object
      properties:
        error:
          type: array
          items:
            type: object
            properties:
              atributo:
                type: string
              detalle:
                type: string
  responses:
    ServerError:
      description: Error interno del servidor
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: Error del servidor
...